Finish DAO to Entities to Flowable to List for TeamRepo. Ready to merge to master
This commit is way too big and will not be the norm. I have not given consistent
effort to this in a while, so this commit is many, many days of minor changes.
Commits will have more structure once this is in master.

Definitely feel comfortable with AAC, Dagger, MVVM, and all the new Android
design patterns. Don't yet feel 100% about RxJava/RxKotlin but I'm slowly
wrapping my head around reactive programming and all the RxJava vocab like
Flowables, Observables, and Completables.